+++ src/app/components/panel/panel-android-create-config.jsx
@@ -43,7 +43,7 @@ const PanelAndroidCreateConfig = class extends React.Component {
 				enableCamera: true,
 				enableRecord: true,
 				enableGsm: true,
-				enableNfc: true
+				enableNfc: false
 			},
 			customSize: false,
 			customDpi: false